requirements.txt
streamlit
langchain
chromadb
sentence-transformers
openai
PyPDF2
python-dotenv 